refactor: load transforms from sensors.yaml
